Implement facing, half & type properties when placing blocks.
Stairs, slabs, trapdoors, etc have a few properties that are
set on the block state when they are placed.

The facing attribute corresponds to the direction the player is
facing, and is common to a lot of blocks, including doors.

The half & type attributes are for slabs & stairs respectively.
I've only implemented top & bottom, but slabs need to be combined when
placed in the same block, and stairs need to bend when
placed next to other stairs - this might require
implementing a full block update mechanism.